git-svn-id: https://svn.viathinksoft.com/svn/filter_foundry/trunk@454 1b5a0048-3058-4200-831f-25aa9e3a14b9

+++ slider_win.c
@@ -105,7 +105,7 @@ Boolean MakeSimpleSubclass(LPCTSTR targetClass, LPCTSTR sourceClass) {
 		if (RegisterClass(&clx) == 0) {
 			TCHAR s[0x300];
 			xstrcpy(s, (TCHAR*)TEXT("RegisterClass failed: "));
-			FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM, NULL, GetLastError(), 0, s + xstrlen(s), 0x300 - xstrlen(s), NULL);
+			FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM, NULL, GetLastError(), 0, s + xstrlen(s), 0x300 - (DWORD)xstrlen(s), NULL);
 			dbg(&s[0]);
 			return false;
 		}
