+++ ui_build_win.c
@@ -72,7 +72,7 @@ Boolean builddialog(FilterRecordPtr pb){
 	if (res == -1) {
 		TCHAR s[0x300];
 		xstrcpy(s, (TCHAR*)TEXT("DialogBoxParam failed: "));
-		FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM, NULL, GetLastError(), 0, s + xstrlen(s), 0x300 - xstrlen(s), NULL);
+		FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM, NULL, GetLastError(), 0, s + xstrlen(s), 0x300 - (DWORD)xstrlen(s), NULL);
 		dbg(&s[0]);
 	}
